Refactor Home Screens
 - Also uses navigator pattern
